PTL/requirements.txt
streamlit
requests
google-cloud-aiplatform
Pillow
matplotlib
python-dotenv
plotly
textblob
pandas
